requirements.txt
torch==2.3.0+cpu
torchvision==0.18.0+cpu
numpy
pandas
Pillow
--extra-index-url https://download.pytorch.org/whl/cpu
